requirements.txt
langchain
streamlit
pypdf2
python-dotenv
faiss-cpu
openai
huggingface_hub
InstructorEmbedding
sentence_transformers
langchain-community